[NEEDS FLAGGING] Extend stmt end position to after semicolon
TODO: we need to fix semicolon after end of do block

Right now, the stmt end position doesn't include the semicolon. We need to extend the position after the semicolon.

We cannot just use end position + 1, because there might be trivia between the end of the stmt and the semicolon