Introduce 'writeMultiline' for StringWriter
Right now, 'write' in StringWriter assumes the string will be on a single line, and hence does not have any multiline logic when updating the position.

This breaks the position tracking when a multiline string is written. This is possible for multiline string literals (or, in future, whitespace and comments)